minor adjustments
Since we are using TMC2209s why not enable 32 microsteps? Adjusted steps_per_unit to comensate for this.

I was also having issues with Z1 not raising/grinding because of all the weight of the controller (and spindle if its cutting on that side). Increased to 900mA from 800mA..

Also enabled square wave stepping for smoother actions. (I definitly have smoother movements)